Move tests to the tests folder
